Moved int test for sendendpoints to pkg
