+++ endpointmanager/cmd/sendendpoints/main.go
@@ -3,11 +3,10 @@ package main
 import (
 	"context"
 	"sync"
-	"time"
 
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/config"
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/endpointmanager/postgresql"
-	"github.com/onc-healthit/lantern-back-end/lanternmq"
+	se "github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/sendendpoints"
 	"github.com/onc-healthit/lantern-back-end/lanternmq/pkg/accessqueue"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
@@ -16,52 +15,6 @@ import (
 func failOnError(err error) {
 	if err != nil {
 		log.Fatalf("%s", err)
-	}
-}
-
-// GetEnptsAndSend gets the current list of endpoints from the database and sends each one to the given queue
-// it continues to repeat this action every time the given interval period has passed
-func GetEnptsAndSend(
-	ctx context.Context,
-	wg *sync.WaitGroup,
-	qName string,
-	qInterval int,
-	store *postgresql.Store,
-	mq *lanternmq.MessageQueue,
-	channelID *lanternmq.ChannelID,
-	errs chan<- error) {
-
-	defer wg.Done()
-
-	for {
-		listOfEndpoints, err := store.GetAllFHIREndpoints(ctx)
-		if err != nil {
-			errs <- err
-		}
-
-		err = accessqueue.SendToQueue(ctx, "start", mq, channelID, qName)
-		if err != nil {
-			errs <- err
-		}
-
-		for i, endpt := range listOfEndpoints {
-			if i%10 == 0 {
-				log.Infof("Processed %d/%d messages", i, len(listOfEndpoints))
-			}
-
-			err = accessqueue.SendToQueue(ctx, endpt.URL, mq, channelID, qName)
-			if err != nil {
-				errs <- err
-			}
-		}
-
-		err = accessqueue.SendToQueue(ctx, "stop", mq, channelID, qName)
-		if err != nil {
-			errs <- err
-		}
-
-		log.Infof("Waiting %d minutes", qInterval)
-		time.Sleep(time.Duration(qInterval) * time.Minute)
 	}
 }
 
@@ -91,11 +44,11 @@ func main() {
 	ctx := context.Background()
 	wg.Add(1)
 	capInterval := viper.GetInt("capquery_qryintvl")
-	go GetEnptsAndSend(ctx, &wg, capQName, capInterval, store, &mq, &channelID, errs)
+	go se.GetEnptsAndSend(ctx, &wg, capQName, capInterval, store, &mq, &channelID, errs)
 
 	wg.Add(1)
 	netInterval := viper.GetInt("endptqry_query_interval")
-	go GetEnptsAndSend(ctx, &wg, netQName, netInterval, store, &mq, &channelID, errs)
+	go se.GetEnptsAndSend(ctx, &wg, netQName, netInterval, store, &mq, &channelID, errs)
 
 	for elem := range errs {
 		log.Warn(elem)
